Moved diagrams from unit test to main.
