+++ src/tw/com/pictures/NetworkChildDiagram.java
@@ -1,7 +1,5 @@
 package tw.com.pictures;
 
-import tw.com.unit.TemplatedChildDiagram;
-
 public class NetworkChildDiagram extends TemplatedChildDiagram<ChildDiagram> {
 
 	public NetworkChildDiagram(ChildDiagram contained) {
