+++ src/tw/com/pictures/TemplatedChildDiagram.java
@@ -1,4 +1,4 @@
-package tw.com.unit;
+package tw.com.pictures;
 
 import tw.com.exceptions.CfnAssistException;
 import tw.com.pictures.ChildDiagram;
